Updated Draw method and added vertical offset
The Draw method has been updated to better support drawable bounds. There are some cases the bounds will be changed before the draw method is called (LayerDrawable) and the size specified is overridden. Also a fix has been applied to the origin calc if the bounds is not a square and taller than the icon.